+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2025-05-23 - Feature: Add course related hint for guest access for teachers as well, resolves #960.
 * 2025-05-23 - Improvement: Re-sort the settings for the course related hints.
 
 ### v4.5-r17